requirements.txt
asgiref==3.2.3
Django==2.1.5
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Markdown==3.2.1
Pillow==7.0.0
PyJWT==1.7.1
pytz==2018.9
sqlparse==0.3.0
